Bug fixing and explainer rename
