+++ pyreal/explainers/__init__.py
@@ -27,8 +27,8 @@ from pyreal.explainers.time_series.saliency.univariate_lime_saliency import (
     UnivariateLimeSaliency,
 )
 
-from pyreal.explainers.example.base import ExampleBasedBase
-from pyreal.explainers.example.similar_examples import SimilarExamples
+from pyreal.explainers.se.base import SimilarExamplesBase
+from pyreal.explainers.se.similar_examples import SimilarExamples
 
 from pyreal.explainers.generic_explainer import Explainer
 
@@ -52,7 +52,7 @@ __all__ = [
     "SaliencyBase",
     "UnivariateOcclusionSaliency",
     "UnivariateLimeSaliency",
-    "ExampleBasedBase",
+    "SimilarExamplesBase",
     "SimilarExamples",
     "Explainer",
 ]
